Removed *free* promise
For communities like warmshowers I'd hope to be for free forever (donations based), as I believe those communities should not charge theirs users either. But we need to earn somehow to pay for infrastructure and maintenance. If donations were not enough I can imagine this could be done by serving money-exchange communities in addition to reciprocal ones, and charging them some fee.