Create LICENSE
MIT because fedi-trustroots is on MIT.

This repo serves as an issue tracker and documentation for fedi-trustroots and OHN in general. When we change fedi-trustroots license we’ll change this one as well.

There is no CC0 or other because per my understanding if the content is domumentation it is not needed. If we started a blog we may add license for that.